+++ internal/adapters/ent/schema/sendchunk.go
@@ -19,7 +19,7 @@ func (SendChunk) Fields() []ent.Field {
 		field.Int("file_id").Unique(),
 		field.Int("chunk_index").Default(0),
 		field.Int64("chunk_offset").Default(0),
-		field.Int64("chunk_size").Default(0),
+		field.Int("chunk_size").Default(0),
 		field.Int("status").Default(0),
 		field.Time("updated_at").Default(time.Now).UpdateDefault(time.Now),
 		field.Time("created_at").Default(time.Now).Immutable(),
@@ -31,7 +31,7 @@ func (SendChunk) Edges() []ent.Edge {
 	return []ent.Edge{
 		edge.From("send_file", SendFile.Type).
 			Ref("send_chunks").
-			Field("file_id").
+			Field("file_id"). // 和sendfile表的id字段关联
 			// 外键关联
 			Unique().
 			Required(),
